# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Skills4CRM API
  version: 0.2.0
  description: Employer Engagement CRM API — Skills4CRM
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List all contacts
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: organisationId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get contact by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          description: Not found
    put:
      operationId: updateContact
      tags: [contacts]
      summary: Update contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContactRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- Engagements ---
  /engagements:
    get:
      operationId: listEngagements
      tags: [engagements]
      summary: List all engagements
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: stage
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: organisationId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of engagements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Engagement"
    post:
      operationId: createEngagement
      tags: [engagements]
      summary: Create engagement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEngagementRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Engagement"
  /engagements/{id}:
    get:
      operationId: getEngagement
      tags: [engagements]
      summary: Get engagement by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Engagement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Engagement"
        "404":
          description: Not found
    put:
      operationId: updateEngagement
      tags: [engagements]
      summary: Update engagement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEngagementRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Engagement"
    delete:
      operationId: deleteEngagement
      tags: [engagements]
      summary: Delete engagement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- Tasks ---
...
components:
  schemas:
...
    PreferredContactMethod:
      type: string
      enum: [email, phone, post, no_preference]
    Contact:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        jobTitle:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        preferredContactMethod:
          $ref: "#/components/schemas/PreferredContactMethod"
        organisationId:
          type: integer
          nullable: true
        organisationName:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, firstName, lastName, createdAt, updatedAt]
    CreateContactRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        jobTitle:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        preferredContactMethod:
          $ref: "#/components/schemas/PreferredContactMethod"
        organisationId:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
      required: [firstName, lastName]
    UpdateContactRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        jobTitle:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        preferredContactMethod:
          $ref: "#/components/schemas/PreferredContactMethod"
        organisationId:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true

    # --- Engagement ---
    EngagementStage:
      type: string
      enum: [lead, contacted, meeting_booked, proposal, active, won, dormant]
    EngagementStatus:
      type: string
      enum: [open, closed_won, closed_lost, on_hold]
    Engagement:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        organisationId:
          type: integer
          nullable: true
        organisationName:
          type: string
          nullable: true
        primaryContactId:
          type: integer
          nullable: true
        contactName:
          type: string
          nullable: true
        ownerUserId:
          type: integer
          nullable: true
        ownerName:
          type: string
          nullable: true
        stage:
          $ref: "#/components/schemas/EngagementStage"
        status:
          $ref: "#/components/schemas/EngagementStatus"
        expectedLearnerVolume:
          type: integer
          nullable: true
        expectedValue:
          type: number
          nullable: true
        probability:
          type: integer
          nullable: true
        lastContactDate:
          type: string
          nullable: true
        nextActionDate:
          type: string
          nullable: true
        nextActionNote:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, stage, status, createdAt, updatedAt]
    CreateEngagementRequest:
      type: object
      properties:
        title:
          type: string
        organisationId:
          type: integer
          nullable: true
        primaryContactId:
          type: integer
          nullable: true
        ownerUserId:
          type: integer
          nullable: true
        stage:
          $ref: "#/components/schemas/EngagementStage"
        status:
          $ref: "#/components/schemas/EngagementStatus"
        expectedLearnerVolume:
          type: integer
          nullable: true
        expectedValue:
          type: number
          nullable: true
        probability:
          type: integer
          nullable: true
        lastContactDate:
          type: string
          nullable: true
        nextActionDate:
          type: string
          nullable: true
        nextActionNote:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [title, stage, status]
    UpdateEngagementRequest:
      type: object
      properties:
        title:
          type: string
        organisationId:
          type: integer
          nullable: true
        primaryContactId:
          type: integer
          nullable: true
        ownerUserId:
          type: integer
          nullable: true
        stage:
          $ref: "#/components/schemas/EngagementStage"
        status:
          $ref: "#/components/schemas/EngagementStatus"
        expectedLearnerVolume:
          type: integer
          nullable: true
        expectedValue:
          type: number
          nullable: true
        probability:
          type: integer
          nullable: true
        lastContactDate:
          type: string
          nullable: true
        nextActionDate:
          type: string
          nullable: true
        nextActionNote:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true

    # --- Task ---
...